Can now change motion-direction in visualiser app
The blue line in the app is the origin, the red line is the current rotation
direction (change with left click), the orange line is the current movement
direction (change with right click). The longer the orange or red line, the
faster the robot is rotating/moving.
The blob in the centre is the robot. The circle around the robot indicates the
maximum speed at which the robot can rotate/travel.

Closes #7